citadel_master/requirements.txt
grpcio==1.73.1
grpcio-tools==1.73.1
grpcio-reflection==1.73.1
protobuf==6.31.0
numpy==2.2.6
watchfiles==1.0.5
colorama>=0.4
psycopg2-binary==2.9.10
orjson==3.11.0